Refactor BaGet (#214)
* Rename `BaGet.Controllers` to `BaGet.Core.Server`
* Move code from `BaGet` to `BaGet.Core` or `BaGet.Core.Server` projects
