+++ src/BaGet.Core.Server/BaGet.Core.Server.csproj
@@ -1,19 +1,18 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;net461</TargetFrameworks>
 
-    <Description>BaGet's ASP.NET Core API controllers</Description>
+    <Description>BaGet's NuGet server implementation</Description>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.HttpOverrides" version="$(MicrosoftAspNetCorePackageVersion)" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc" version="$(MicrosoftAspNetCorePackageVersion)" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\BaGet.Core\BaGet.Core.csproj" />
-    <ProjectReference Include="..\BaGet.AWS\BaGet.AWS.csproj" />
-    <ProjectReference Include="..\BaGet.Azure\BaGet.Azure.csproj" />
   </ItemGroup>
 </Project>
