+++ src/BaGet.Core.Server/Configuration/ConfigureForwardedHeadersOptions.cs
@@ -1,0 +1,18 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.HttpOverrides;
+using Microsoft.Extensions.Options;
+
+namespace BaGet.Configuration
+{
+    public class ConfigureForwardedHeadersOptions : IConfigureOptions<ForwardedHeadersOptions>
+    {
+        public void Configure(ForwardedHeadersOptions options)
+        {
+            options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+
+            // Do not restrict to local network/proxy
+            options.KnownNetworks.Clear();
+            options.KnownProxies.Clear();
+        }
+    }
+}
