+++ src/BaGet.Core.Server/Controllers/Registration/RegistrationIndexController.cs
