+++ src/BaGet.Core.Server/Extensions/UrlExtensions.cs
