+++ src/readme.md
@@ -2,10 +2,10 @@
 
 These folders contain the core components of BaGet:
 
-* `BaGet` - the app's entry point and CLI commands
-* `BaGet.Controllers` - the  API controllers
+* `BaGet` - the app's entry point that glues everything together
 * `BaGet.Core` - BaGet's core logic and services
-* `BaGet.Protocol` - Libraries to interact with NuGet servers' APIs
+* `BaGet.Core.Server` - The services that implement [the NuGet server APIs](https://docs.microsoft.com/en-us/nuget/api/overview) using `BaGet.Core`
+* `BaGet.Protocol` - Libraries to interact with [NuGet servers' APIs](https://docs.microsoft.com/en-us/nuget/api/overview)
 * `BaGet.UI` - BaGet's React frontend
 
 These folders contain cloud-specific components of BaGet:
